Remove redundant launch file
I noticed that this launch file was still lying around from #212. I had
added it in an early commit of #212 as we initially did not want the
reservation system running in the default world. However, now that the
reservation system is running in the default office world, this launch file is
redundant. Additionally, this launch file references a non-existant
launch file in rmf_demos and hence will not work. To reduce confusion
I am removing this launch file.

Signed-off-by: Arjo Chakravarty <arjoc@google.com>